+++ packages/evokit-flex/__tests__/Flex.content.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import barista from 'seed-barista';
+import style from '../style.css';
 import { Flex } from '../src';
 
 describe('<Flex flex-content="..." />', () => {
@@ -29,5 +31,30 @@ describe('<Flex flex-content="..." />', () => {
         expect(wrapper).toMatchSnapshot();
 
         expect(wrapper.html()).toEqual('<div class="ek-flex ek-flex_content-justify_end"></div>');
+    });
+
+    it('css', () => {
+        const css = barista({ content: style });
+
+        expect(css.rule('.ek-flex_content_start').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_end').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_center').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_between').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_around').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_stretch').exists()).toBeTruthy();
+
+        expect(css.rule('.ek-flex_content_align-start').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_align-end').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_align-center').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_align-between').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_align-around').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_align-stretch').exists()).toBeTruthy();
+
+        expect(css.rule('.ek-flex_content_justify-start').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_justify-end').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_justify-center').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_justify-between').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_justify-around').exists()).toBeTruthy();
+        expect(css.rule('.ek-flex_content_justify-stretch').exists()).toBeTruthy();
     });
 });